Extract submission flow into src/core/submission.ts
Move verification submission, user grading/scoring, content-creator
detection, and pending account snapshot collection into a dedicated
module. Typecheck, 167 tests, and build all pass.
